+++ src/components/Search.vue
@@ -8,7 +8,7 @@ const searchStore = useSearchStore();
     <el-input
       v-model="searchStore.input"
       v-show="searchStore.visible"
-      placeholder="Please input for search"
+      placeholder="Please input keywords for searching in this page"
       class="w-full search-input absolute bottom-0 font-mono font-bold z-50"
       size="large"
       @click="searchStore.search"